+++ src/features/auth/hooks/useNicknameSetupForm.ts
@@ -1,17 +1,13 @@
 import { FormEvent, useEffect, useMemo, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useRequireNicknameSetupSession } from './useRequireNicknameSetupSession'
-import {
-  mockCheckNicknameAvailability,
-  mockSetNickname,
-  validateNickname,
-} from '../mockApi'
+import { mockSetNickname } from '../mockApi'
 import {
   createNicknameHelperFeedback,
-  NICKNAME_CHECK_DEBOUNCE_MS,
   VALIDATION_MESSAGE_DEBOUNCE_MS,
 } from '../nicknameSetupRules'
 import { useAuthStore } from '../store'
+import { useNicknameAvailability } from './useNicknameAvailability'
 
 // 닉네임 설정 화면의 입력/검증/제출 상태를 통합 관리
 export function useNicknameSetupForm() {
@@ -24,54 +20,10 @@ export function useNicknameSetupForm() {
   const [submitMessage, setSubmitMessage] = useState<string | null>(null)
   const [isSubmitting, setIsSubmitting] = useState(false)
   const [isNicknameFocused, setIsNicknameFocused] = useState(false)
-  const [isCheckingNickname, setIsCheckingNickname] = useState(false)
-  const [isNicknameAvailable, setIsNicknameAvailable] = useState<
-    boolean | null
-  >(null)
   const [showValidationMessage, setShowValidationMessage] = useState(false)
 
-  // 입력값이 바뀔 때마다 로컬 형식 검증 결과 계산
-  const nicknameValidation = useMemo(
-    () => validateNickname(nickname),
-    [nickname]
-  )
-
-  // 형식 검증 통과 시 디바운스로 닉네임 중복 검사 실행
-  useEffect(() => {
-    setIsCheckingNickname(false)
-    setIsNicknameAvailable(null)
-
-    // 형식 오류가 있으면 중복 검사 생략
-    if (!nicknameValidation.ok) {
-      return
-    }
-
-    let isDisposed = false
-    const normalizedNickname = nicknameValidation.normalizedNickname
-    const debounceTimer = window.setTimeout(async () => {
-      setIsCheckingNickname(true)
-      const result = await mockCheckNicknameAvailability(normalizedNickname)
-      // 언마운트/재실행 이후 응답은 상태 반영 차단
-      if (isDisposed) {
-        return
-      }
-
-      setIsCheckingNickname(false)
-      // API가 실패를 반환하면 사용 가능 여부 초기화
-      if (!result.ok) {
-        setIsNicknameAvailable(null)
-        return
-      }
-
-      setIsNicknameAvailable(result.isAvailable)
-    }, NICKNAME_CHECK_DEBOUNCE_MS)
-
-    // 디바운스 타이머 정리와 비동기 응답 무시 처리
-    return () => {
-      isDisposed = true
-      window.clearTimeout(debounceTimer)
-    }
-  }, [nicknameValidation])
+  const { nicknameValidation, isCheckingNickname, isNicknameAvailable } =
+    useNicknameAvailability(nickname)
 
   // 입력 직후 안내 메시지 노출을 지연해 깜빡임 완화
   useEffect(() => {
